+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - added config option for space station maximum size
 - added config option for rocket crash explosions
 - added GT 1.3.1 dimension marker blocks (no textures yet though, sorry!)
+- added text to space station packager GUI
 
 ### FIXES:
 - fixed the rocket scanner not working when facing north/south